Moved normalizer to its own package
